Using toast for service worker prompt
