+++ src/components/Toast/index.tsx
@@ -16,7 +16,7 @@ const Toast: React.FC<ToastProps> = ({ children, onDismiss }) => {
   return (
     <div
       className={cn(
-        'px-4 py-2 bg-gray-700 rounded-sm fixed right-10 bottom-10 md:min-w-toast',
+        'px-4 py-2 bg-gray-700 rounded-sm fixed w-toast bottom-16 sm:w-auto md:right-10 md:bottom-10 md:min-w-toast',
         animationState === 'entry' && 'animate-toast-entry',
         animationState === 'exit' && 'animate-toast-exit'
       )}
